boards: rcar_ebisu: do not hijack the console
On Ebisu there is only one available UART,
before using it, make sure that Linux or whatever other
component is not using it. Since it's a common scheme to have it
enabled in Linux disable it by default and prefer usage of ram_console
which is available for the remoteproc subsystem.

However the scif2 is possible to use from Zephyr, and definetely helpful
at least when debugging.
Enable it if you know how to handle side effect.

Signed-off-by: Julien Massot <julien.massot@iot.bzh>