drivers: pinmux: Add Renesas R-Car driver
On R-Car setting a function of a multiplexed pin boils down to
two steps:

- Set the matching pin bit int the correct,
GPSR (GPIO/Peripheral Function Select Register) to mark the pin
as GPIO or Peripheral: 1 for peripheral 0 for GPIO.

- Then in case of peripheral function, the function should be set
in the IPSR register (Peripheral Function Select Register).
Each possible function is coded on 4 bits. One 32 bits register
can hold up to 8 pins function.

Since two registers have to be set (IPSR and GPSR) to assign an
alternate function to a pin, the pinmux_driver_api is not used.
Instead a pinmux_rcar_set_pingroup function is available, and allow
to select function for a set of pins, scif1_data_a for example.

Signed-off-by: Julien Massot <julien.massot@iot.bzh>