tests: Bluetooth: BSIM: ISO: Refactor TX
Refactor the TX for the ISO host BSIM tests to use
a separate thread, instead of a timer.
This is the recommended way to handle TX, and is what the
LE Audio BSIM tests and samples use.

The TX will send 1->sdu number of octets, increasing, to
verify that any size can be sent (and to keep the check
for this in the bis_receiver).

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>